+++ src/components/Pdf/PdfDocumentViewer.tsx
@@ -1,11 +1,12 @@
 import { PDFViewer } from '@react-pdf/renderer';
 import { useFormContext } from 'react-hook-form';
 
-import { type ResumeFormSchema } from '../Providers/FormProvider';
+import { type ResumeFormDataSchema } from '@/types/form';
+
 import PdfDocument from './PdfDocument';
 
 const PdfDocumentViewer = () => {
-  const { getValues } = useFormContext<ResumeFormSchema>();
+  const { getValues } = useFormContext<ResumeFormDataSchema>();
   return (
     <PDFViewer width="100%" height="650px" showToolbar={false}>
       <PdfDocument resume={getValues()} />
