--- src/types/resume.ts
@@ -1,7 +1,0 @@
-import { type ActivitiesFormSchema } from '@/components/Form/ActivityForm';
-import { type ProjectsFormSchema } from '@/components/Form/ProjectForm';
-
-export interface Resume {
-  projects: ProjectsFormSchema;
-  activities: ActivitiesFormSchema;
-}
